마나 추가 및 스태미나 버그 수정

+++ Assets/Devs/Dongjun/UI/Player Stat UI/UI_PlayerHealthBar.cs
@@ -13,6 +13,6 @@ public class UI_PlayerHealthBar : MonoBehaviour
 
     private void Update_HealthBar()
     {
-        healthBar.Value = (float)PlayerStats.Inst.health.Value / PlayerStats.Inst.health.Max;
+        healthBar.Value = PlayerStats.Inst.health.Value / PlayerStats.Inst.health.Max;
     }
 }
